Добавил тип 'module' main.js
